backend/requirements.txt
fastapi
uvicorn
sqlalchemy~=2.0
strawberry-graphql[fastapi,debug-server]
